test-prosa/requirements.txt
accelerate==0.27.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b3
azure-core==1.30.2
Babel==2.8.0
backcall==0.2.0
backoff==2.2.1
bcc==0.18.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bitsandbytes==0.43.1
bleach==6.1.0
blessed==1.20.0
blinker==1.8.2
branca==0.7.1
Brlapi==0.8.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.0
chrome-gnome-shell==0.0.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.2.0
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.18.0
dbus-python==1.2.18
debugpy==1.8.2
decorator==5.1.1
deep-translator==1.11.4
deepdiff==7.0.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.8
discord==2.3.2
discord.py==2.4.0
diskcache==5.6.3
distlib==0.3.4
distro==1.9.0
docopt==0.6.2
einops==0.7.0
emoji==2.12.1
environs==9.5.0
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.0
fastjsonschema==2.19.1
filelock==3.15.4
filetype==1.2.0
fiona==1.9.6
flash-attn==2.5.6
Flask==3.0.3
folium==0.16.0
fontawesomefree==6.5.1
fonttools==4.49.0
frozenlist==1.5.0
fsspec==2024.6.1
geopandas==0.14.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gpt4all==2.2.1.post1
gpustat==1.1.1
greenlet==3.1.1
groq==0.11.0
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
h2==4.1.0
hidpidaemon==18.4.6
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.5
hyperframe==6.0.1
idna==3.10
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.2
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.30
langchain-google-community==1.0.7
langchain-groq==0.1.9
langchain-milvus==0.1.6
langchain-openai==0.1.21
langchain-qdrant==0.1.3
langchain-text-splitters==0.2.2
langchain-upstage==0.1.7
langchainhub==0.1.14
langdetect==1.0.9
langserve==0.2.2
langsmith==0.1.95
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama_cpp_python==0.2.55
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.10
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ollama==0.1.7
openai==1.52.2
openpyxl==3.1.5
ordered-set==4.1.0
orjson==3.10.10
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pdfminer.six==20231228
pdfplumber==0.11.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pipreqsnb==0.2.4
platformdirs==4.2.2
plotly==5.20.0
pop-transition==1.1.2
portalocker==2.10.1
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
proton-core==0.2.0
proton-keyring-linux==0.0.2
proton-keyring-linux-secretservice==0.0.2
proton-vpn-api-core==0.29.4
proton-vpn-gtk-app==4.4.2
proton-vpn-killswitch-network-manager==0.5.0
proton-vpn-killswitch-network-manager-wireguard==0.1.0
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.5.0
proton-vpn-network-manager-openvpn==0.0.7
proton-vpn-network-manager-wireguard==0.0.12
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.9.2
pydantic-settings==2.4.0
pydantic_core==2.23.4
pydbus==0.6.0
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymilvus==2.4.9
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.2.0
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproj==3.6.1
pyproject-toml==0.0.10
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-iso639==2024.4.27
python-magic==0.4.27
python-xlib==0.29
pytz==2024.1
pywaffle==1.1.0
pyxdg==0.27
PyYAML==6.0.2
pyzmq==26.0.3
qdrant-client==1.12.0
rapidfuzz==3.9.6
redis==5.0.8
referencing==0.35.1
regex==2024.9.11
repolib==2.2.1
repoman==1.4.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.18.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
sentry-sdk==1.4.3
sessioninstaller==0.0.0
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.36
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.1
streamlit==1.37.0
streamlit_mic_recorder==0.0.8
sympy==1.13.1
systemd-python==234
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.0
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.43.3
triton==3.0.0
types-requests==2.31.0.20240218
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
unstructured==0.15.1
unstructured-client==0.25.5
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.5
virtualenv==20.13.0+ds
wadllib==1.3.6
watchdog==4.0.1
Wave==0.0.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.4
widgetsnbextension==4.0.10
wrapt==1.16.0
xdg==5
xformers==0.0.27.post2
xkit==0.0.0
XlsxWriter==3.2.0
xlwings==0.31.3
xxhash==3.4.1
xyzservices==2023.10.1
yarg==0.1.9
yarl==1.17.0
zipp==1.0.0
